+++ chaosreliably/__init__.py
@@ -13,7 +13,7 @@ from logzero import logger
 __version__ = "0.2.1"
 __all__ = ["get_session", "discover"]
 RELIABLY_CONFIG_PATH = "~/.config/reliably/config.yaml"
-RELIABLY_HOST = "api.reliably.com"
+RELIABLY_HOST = "reliably.com"
 
 
 @contextmanager